Cookieに関するファイルを分離
Cookieに関するファイルを分離。合わせて、require()をブラウザ環境でも（無理やり）共存させた。
